feat(second): implement figure selection feature
Added a new feature to allow users to select a premade figure to place on
the canvas. Added a new text field to the view and a corresponding action
listener in the controller. When the user enters a number in the text
field and presses enter, the corresponding premade figure is set as the
current figure in the model. The highest x and y points of the figure are
calculated to facilitate proper placement. (Denis)